pyproject.toml
[project]
name = "pulmoscanx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "efficientnet-pytorch>=0.7.1",
    "kornia>=0.8.1",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "scikit-learn>=1.7.0",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
